integration: move silent_prepare_query() to unprepared.rs
The test is strictly about unprepared statements, so it's moved to
the corresponding module.
